fix(product): change from hard delete to soft delete for future proof error

+++ src/main/java/com/learningmat/ecommerce/dto/request/ProductRequest.java
@@ -10,7 +10,7 @@ public record ProductRequest(
         String name,
 
         @Min(value = 0, message = "INVALID_PRICE")
-        double price,
+        long price,
 
         Long categoryId) {
 }
